Implement login screen with email and password authentication isolated
