+++ app/src/main/java/com/sedilant/yambol/ui/profile/ProflieScreen.kt
@@ -1,6 +1,5 @@
 package com.sedilant.yambol.ui.profile
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -22,9 +21,7 @@ import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Email
 import androidx.compose.material.icons.filled.Error
-import androidx.compose.material.icons.filled.MarkEmailRead
 import androidx.compose.material.icons.filled.Person
-import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material.icons.filled.Warning
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
@@ -42,26 +39,16 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.credentials.CredentialManager
-import androidx.credentials.GetCredentialRequest
-import androidx.credentials.exceptions.GetCredentialCancellationException
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.sedilant.yambol.R
-import com.google.android.libraries.identity.googleid.GetGoogleIdOption
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
-import kotlinx.coroutines.launch
-
-private const val WEB_CLIENT_ID = "448568418891-4vq7apk7c16o5s69qg6gvns664oc8cif.apps.googleusercontent.com"
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -69,97 +56,25 @@ fun ProfileScreen(
     viewModel: ProfileViewModel = hiltViewModel()
 ) {
     val uiState by viewModel.uiState.collectAsState()
-    val context = LocalContext.current
-    val coroutineScope = rememberCoroutineScope()
-
-    val onGoogleSignIn: () -> Unit = {
-        coroutineScope.launch {
-            try {
-                val credentialManager = CredentialManager.create(context)
-
-                val googleIdOption = GetGoogleIdOption.Builder()
-                    .setFilterByAuthorizedAccounts(false)
-                    .setServerClientId(WEB_CLIENT_ID)
-                    .build()
-
-                val request = GetCredentialRequest.Builder()
-                    .addCredentialOption(googleIdOption)
-                    .build()
-
-                val result = credentialManager.getCredential(
-                    request = request,
-                    context = context
-                )
-
-                val credential = result.credential
-                val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
-                val idToken = googleIdTokenCredential.idToken
-
-                viewModel.signInWithGoogle(idToken)
-            } catch (e: GetCredentialCancellationException) {
-                // User cancelled
-                Log.d("ProfileScreen", "Google Sign-In cancelled by user")
-            } catch (e: Exception) {
-                Log.e("ProfileScreen", "Google Sign-In failed", e)
-                viewModel.onGoogleSignInFailed(e.message ?: "Google Sign-In failed")
-            }
-        }
-    }
 
     ProfileScreenStateless(
         uiState = uiState,
-        onSignInEmailChange = viewModel::onSignInEmailChange,
-        onSignInPasswordChange = viewModel::onSignInPasswordChange,
-        onToggleSignInPasswordVisibility = viewModel::toggleSignInPasswordVisibility,
-        onContinueWithEmail = viewModel::continueWithEmail,
-        onSignIn = viewModel::signIn,
-        onGoBackToEmail = viewModel::goBackToEmail,
-        onSignUpEmailChange = viewModel::onSignUpEmailChange,
-        onSignUpPasswordChange = viewModel::onSignUpPasswordChange,
-        onSignUpConfirmPasswordChange = viewModel::onSignUpConfirmPasswordChange,
-        onToggleSignUpPasswordVisibility = viewModel::toggleSignUpPasswordVisibility,
-        onToggleSignUpConfirmPasswordVisibility = viewModel::toggleSignUpConfirmPasswordVisibility,
-        onSignUp = viewModel::signUp,
-        onNavigateToSignUp = viewModel::navigateToSignUp,
-        onNavigateToSignIn = viewModel::navigateToSignIn,
-        onGoogleSignIn = onGoogleSignIn,
         onSignOut = viewModel::signOut,
         onShowDeleteConfirmation = viewModel::showDeleteConfirmation,
         onHideDeleteConfirmation = viewModel::hideDeleteConfirmation,
         onConfirmDelete = viewModel::deleteAccount,
-        onErrorDismiss = viewModel::clearError,
-        onResendVerification = viewModel::resendVerificationEmail,
-        onCheckVerification = viewModel::checkEmailVerification,
-        onCancelVerification = viewModel::cancelVerificationAndSignOut
+        onErrorDismiss = viewModel::clearError
     )
 }
 
 @Composable
 private fun ProfileScreenStateless(
     uiState: ProfileUiState,
-    onSignInEmailChange: (String) -> Unit,
-    onSignInPasswordChange: (String) -> Unit,
-    onToggleSignInPasswordVisibility: () -> Unit,
-    onContinueWithEmail: () -> Unit,
-    onSignIn: () -> Unit,
-    onGoBackToEmail: () -> Unit,
-    onSignUpEmailChange: (String) -> Unit,
-    onSignUpPasswordChange: (String) -> Unit,
-    onSignUpConfirmPasswordChange: (String) -> Unit,
-    onToggleSignUpPasswordVisibility: () -> Unit,
-    onToggleSignUpConfirmPasswordVisibility: () -> Unit,
-    onSignUp: () -> Unit,
-    onNavigateToSignUp: () -> Unit,
-    onNavigateToSignIn: () -> Unit,
-    onGoogleSignIn: () -> Unit,
     onSignOut: () -> Unit,
     onShowDeleteConfirmation: () -> Unit,
     onHideDeleteConfirmation: () -> Unit,
     onConfirmDelete: () -> Unit,
-    onErrorDismiss: () -> Unit,
-    onResendVerification: () -> Unit,
-    onCheckVerification: () -> Unit,
-    onCancelVerification: () -> Unit
+    onErrorDismiss: () -> Unit
 ) {
     Scaffold { paddingValues ->
         Box(
@@ -188,60 +103,6 @@ private fun ProfileScreenStateless(
                             onDismiss = onHideDeleteConfirmation
                         )
                     }
-                }
-
-                is ProfileUiState.SignIn -> {
-                    SignInContent(
-                        email = uiState.emailInput,
-                        password = uiState.passwordInput,
-                        showPasswordField = uiState.showPasswordField,
-                        isPasswordVisible = uiState.isPasswordVisible,
-                        emailError = uiState.emailError,
-                        passwordError = uiState.passwordError,
-                        isGoogleSignInLoading = uiState.isGoogleSignInLoading,
-                        onEmailChange = onSignInEmailChange,
-                        onPasswordChange = onSignInPasswordChange,
-                        onTogglePasswordVisibility = onToggleSignInPasswordVisibility,
-                        onContinueWithEmail = onContinueWithEmail,
-                        onSignIn = onSignIn,
-                        onGoBackToEmail = onGoBackToEmail,
-                        onNavigateToSignUp = onNavigateToSignUp,
-                        onGoogleSignInClick = onGoogleSignIn
-                    )
-                }
-
-                is ProfileUiState.SignUp -> {
-                    SignUpContent(
-                        email = uiState.emailInput,
-                        password = uiState.passwordInput,
-                        confirmPassword = uiState.confirmPasswordInput,
-                        isPasswordVisible = uiState.isPasswordVisible,
-                        isConfirmPasswordVisible = uiState.isConfirmPasswordVisible,
-                        emailError = uiState.emailError,
-                        passwordValidation = uiState.passwordValidation,
-                        confirmPasswordError = uiState.confirmPasswordError,
-                        isGoogleSignInLoading = uiState.isGoogleSignInLoading,
-                        onEmailChange = onSignUpEmailChange,
-                        onPasswordChange = onSignUpPasswordChange,
-                        onConfirmPasswordChange = onSignUpConfirmPasswordChange,
-                        onTogglePasswordVisibility = onToggleSignUpPasswordVisibility,
-                        onToggleConfirmPasswordVisibility = onToggleSignUpConfirmPasswordVisibility,
-                        onSignUp = onSignUp,
-                        onNavigateToSignIn = onNavigateToSignIn,
-                        onGoogleSignInClick = onGoogleSignIn
-                    )
-                }
-
-                is ProfileUiState.EmailVerificationPending -> {
-                    EmailVerificationContent(
-                        email = uiState.email,
-                        isResending = uiState.isResending,
-                        isChecking = uiState.isChecking,
-                        message = uiState.message,
-                        onResendClick = onResendVerification,
-                        onCheckVerificationClick = onCheckVerification,
-                        onCancelClick = onCancelVerification
-                    )
                 }
 
                 is ProfileUiState.Error -> {
@@ -484,399 +345,6 @@ private fun DeleteAccountDialog(
     )
 }
 
-@Composable
-private fun EmailVerificationContent(
-    email: String,
-    isResending: Boolean,
-    isChecking: Boolean,
-    message: String?,
-    onResendClick: () -> Unit,
-    onCheckVerificationClick: () -> Unit,
-    onCancelClick: () -> Unit
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .verticalScroll(rememberScrollState())
-            .padding(24.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Box(
-            modifier = Modifier
-                .size(120.dp)
-                .clip(CircleShape)
-                .background(MaterialTheme.colorScheme.primaryContainer),
-            contentAlignment = Alignment.Center
-        ) {
-            Icon(
-                imageVector = Icons.Default.MarkEmailRead,
-                contentDescription = stringResource(R.string.cd_email_verification),
-                modifier = Modifier.size(64.dp),
-                tint = MaterialTheme.colorScheme.onPrimaryContainer
-            )
-        }
-
-        Spacer(modifier = Modifier.height(32.dp))
-
-        Text(
-            text = stringResource(R.string.verify_your_email),
-            style = MaterialTheme.typography.headlineMedium,
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colorScheme.onSurface
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            text = stringResource(R.string.verification_sent_to),
-            style = MaterialTheme.typography.bodyLarge,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            text = email,
-            style = MaterialTheme.typography.bodyLarge,
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colorScheme.primary
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            text = stringResource(R.string.verification_check_inbox),
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.padding(horizontal = 16.dp)
-        )
-
-        message?.let {
-            Spacer(modifier = Modifier.height(16.dp))
-            Card(
-                modifier = Modifier.fillMaxWidth(),
-                colors = CardDefaults.cardColors(
-                    containerColor = if (it.contains("sent", ignoreCase = true))
-                        MaterialTheme.colorScheme.primaryContainer
-                    else
-                        MaterialTheme.colorScheme.errorContainer
-                )
-            ) {
-                Row(
-                    modifier = Modifier.padding(16.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Icon(
-                        imageVector = if (it.contains("sent", ignoreCase = true))
-                            Icons.Default.CheckCircle
-                        else
-                            Icons.Default.Warning,
-                        contentDescription = null,
-                        tint = if (it.contains("sent", ignoreCase = true))
-                            MaterialTheme.colorScheme.primary
-                        else
-                            MaterialTheme.colorScheme.error
-                    )
-                    Spacer(modifier = Modifier.width(12.dp))
-                    Text(
-                        text = it,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = if (it.contains("sent", ignoreCase = true))
-                            MaterialTheme.colorScheme.onPrimaryContainer
-                        else
-                            MaterialTheme.colorScheme.onErrorContainer
-                    )
-                }
-            }
-        }
-
-        Spacer(modifier = Modifier.height(32.dp))
-
-        Button(
-            onClick = onCheckVerificationClick,
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(56.dp),
-            enabled = !isChecking && !isResending,
-            shape = MaterialTheme.shapes.medium
-        ) {
-            if (isChecking) {
-                CircularProgressIndicator(
-                    modifier = Modifier.size(24.dp),
-                    color = MaterialTheme.colorScheme.onPrimary,
-                    strokeWidth = 2.dp
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-            }
-            Text(
-                text = if (isChecking) stringResource(R.string.checking) else stringResource(R.string.verified_my_email),
-                style = MaterialTheme.typography.titleMedium
-            )
-        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        OutlinedButton(
-            onClick = onResendClick,
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(56.dp),
-            enabled = !isResending && !isChecking,
-            shape = MaterialTheme.shapes.medium
-        ) {
-            if (isResending) {
-                CircularProgressIndicator(
-                    modifier = Modifier.size(24.dp),
-                    strokeWidth = 2.dp
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-            } else {
-                Icon(
-                    imageVector = Icons.Default.Refresh,
-                    contentDescription = null,
-                    modifier = Modifier.size(20.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-            }
-            Text(
-                text = if (isResending) stringResource(R.string.sending) else stringResource(R.string.resend_verification_email),
-                style = MaterialTheme.typography.titleMedium
-            )
-        }
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        TextButton(
-            onClick = onCancelClick,
-            enabled = !isResending && !isChecking
-        ) {
-            Text(
-                text = stringResource(R.string.cancel_and_sign_out),
-                color = MaterialTheme.colorScheme.error
-            )
-        }
-    }
-}
-
-// ======================= PREVIEWS =======================
-
-@Preview(showBackground = true, name = "Loading State")
-@Composable
-private fun ProfileScreenLoadingPreview() {
-    ProfileScreenStateless(
-        uiState = ProfileUiState.Loading,
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
-        onSignOut = {},
-        onShowDeleteConfirmation = {},
-        onHideDeleteConfirmation = {},
-        onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
-    )
-}
-
-@Preview(showBackground = true, name = "Sign In - Email Step")
-@Composable
-private fun ProfileScreenSignInEmailPreview() {
-    ProfileScreenStateless(
-        uiState = ProfileUiState.SignIn(
-            emailInput = "user@example.com",
-            showPasswordField = false
-        ),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
-        onSignOut = {},
-        onShowDeleteConfirmation = {},
-        onHideDeleteConfirmation = {},
-        onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
-    )
-}
-
-@Preview(showBackground = true, name = "Sign In - Password Step")
-@Composable
-private fun ProfileScreenSignInPasswordPreview() {
-    ProfileScreenStateless(
-        uiState = ProfileUiState.SignIn(
-            emailInput = "user@example.com",
-            showPasswordField = true
-        ),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
-        onSignOut = {},
-        onShowDeleteConfirmation = {},
-        onHideDeleteConfirmation = {},
-        onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
-    )
-}
-
-@Preview(showBackground = true, name = "Sign Up")
-@Composable
-private fun ProfileScreenSignUpPreview() {
-    ProfileScreenStateless(
-        uiState = ProfileUiState.SignUp(
-            emailInput = "newuser@example.com",
-            passwordInput = "pass1!",
-            confirmPasswordInput = "",
-            passwordValidation = PasswordValidation(
-                isValid = true,
-                hasMinLength = true,
-                hasSymbol = true,
-                hasNumber = true
-            )
-        ),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
-        onSignOut = {},
-        onShowDeleteConfirmation = {},
-        onHideDeleteConfirmation = {},
-        onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
-    )
-}
-
-@Preview(showBackground = true, name = "Sign Up - Password Validation")
-@Composable
-private fun ProfileScreenSignUpValidationPreview() {
-    ProfileScreenStateless(
-        uiState = ProfileUiState.SignUp(
-            emailInput = "newuser@example.com",
-            passwordInput = "pass",
-            confirmPasswordInput = "",
-            passwordValidation = PasswordValidation(
-                isValid = false,
-                hasMinLength = false,
-                hasSymbol = false,
-                hasNumber = false
-            )
-        ),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
-        onSignOut = {},
-        onShowDeleteConfirmation = {},
-        onHideDeleteConfirmation = {},
-        onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
-    )
-}
-
-@Preview(showBackground = true, name = "Email Verification Pending")
-@Composable
-private fun ProfileScreenEmailVerificationPreview() {
-    ProfileScreenStateless(
-        uiState = ProfileUiState.EmailVerificationPending(
-            email = "newuser@example.com",
-            verificationSent = true,
-            message = "Verification email sent! Please check your inbox."
-        ),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
-        onSignOut = {},
-        onShowDeleteConfirmation = {},
-        onHideDeleteConfirmation = {},
-        onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
-    )
-}
-
 @Preview(showBackground = true, name = "Authenticated")
 @Composable
 private fun ProfileScreenAuthenticatedPreview() {
@@ -888,29 +356,11 @@ private fun ProfileScreenAuthenticatedPreview() {
             photoUrl = null,
             showDeleteConfirmation = false
         ),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
         onSignOut = {},
         onShowDeleteConfirmation = {},
         onHideDeleteConfirmation = {},
         onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
+        onErrorDismiss = {}
     )
 }
 
@@ -919,33 +369,10 @@ private fun ProfileScreenAuthenticatedPreview() {
 private fun ProfileScreenErrorPreview() {
     ProfileScreenStateless(
         uiState = ProfileUiState.Error(message = "Authentication failed. Please check your credentials."),
-        onSignInEmailChange = {},
-        onSignInPasswordChange = {},
-        onToggleSignInPasswordVisibility = {},
-        onContinueWithEmail = {},
-        onSignIn = {},
-        onGoBackToEmail = {},
-        onSignUpEmailChange = {},
-        onSignUpPasswordChange = {},
-        onSignUpConfirmPasswordChange = {},
-        onToggleSignUpPasswordVisibility = {},
-        onToggleSignUpConfirmPasswordVisibility = {},
-        onSignUp = {},
-        onNavigateToSignUp = {},
-        onNavigateToSignIn = {},
-        onGoogleSignIn = {},
         onSignOut = {},
         onShowDeleteConfirmation = {},
         onHideDeleteConfirmation = {},
         onConfirmDelete = {},
-        onErrorDismiss = {},
-        onResendVerification = {},
-        onCheckVerification = {},
-        onCancelVerification = {}
+        onErrorDismiss = {}
     )
 }
-
-
-
-
-
